Fixed a bug with empty events (#45)
* Fixed a bug with empty events

When an event was empty (i.e. no object, no summary, almost nothing but an id, uid), the event parsing lead to a bug that would cause the application to fail loading the table of people or locations. This was mainly due to the start and end dates not being declared as possibly null when they should have been.
I also cleaned up some code with array accesses that would spam a lot of errors in the Nextcloud logging because the offsets do not exist. I just used null-safe operators to set default values when this happens.

* v0.0.37

---------

Co-authored-by: Matthieu Renard <m.renard@adacis.net>
Co-authored-by: Benjamin <benjamin@cybercorp.fr>